ASoC: SOF: add probe_complete callback definition
With the device model, there is no guarantee that the driver probe of
the child platform device will complete synchronously. The errors are
also squelched so that e.g. the pci driver has no way of knowing that
the firmware boot failed.

Also the probe might be implemented in a workqueue (e.g. if there are
any dependencies on modules or long hardware initialization times), so
we really need an asynchronous way of notifying the parent device that
all probe-related functionality completed successfully in the
"sof-audio" device.

The simplest possible implementation is a callback added to the
sof_pdata structure.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit fa037413506b42bb3b9145dc45bb24737b329565)