ASoC: SOF: debug: specify debugfs memory access type
This patch does the following:
1. Add an access_type member to the debugfs map and the
dfsentry structure if the debugfs memory is always accessible
or only if the DSP is in D0.
2. Update the access_type in the debug_fs map for all
platforms.
The assumption is:
For SKL- platforms, iram/dram and memory windows will be
accessible only in D0.
For SKL+ platforms all memory windows will be accessible
only on D0.
2. If the memory is accessible only in D0, then create
a cache buffer in the debugfs entry that will be used
to copy the contents prior to the DSP entering d3.

3. replace snd_sof_debugfs_io_create_item() with
snd_sof_debugfs_io_item() and
snd_sof_debugfs_buf_create_item with
snd_sof_debugfs_buf_item.
This will help with avoiding checkpatch errors with
the lines shooting over 80 chars.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>