ASoC: SOF: bind pcm to host comp when the dai is loaded
Previously, PCM's were bound to the host component by
using an explicit pipeline graph path from the PCM device
to the host comp in topology. This graph route is unnecessary
and it results in a duplicate connection shown for the host
dapm widget from the PCM device. Instead, pcm's should be bound
to the respective host comp when the dai is loaded.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>