UPSTREAM: ALSA: info: Fix racy addition/deletion of nodes
The ALSA proc helper manages the child nodes in a linked list, but its
addition and deletion is done without any lock.  This leads to a
corruption if they are operated concurrently.  Usually this isn't a
problem because the proc entries are added sequentially in the driver
probe procedure itself.  But the card registrations are done often
asynchronously, and the crash could be actually reproduced with
syzkaller.

This patch papers over it by protecting the link addition and deletion
with the parent's mutex.  There is "access" mutex that is used for the
file access, and this can be reused for this purpose as well.

Reported-by: syzbot+48df349490c36f9f54ab@syzkaller.appspotmail.com
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 8c2f870890fd28e023b0fcf49dcee333f2c8bad7)

BUG=b:134688240
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: I767c568d1737eceef0c9bd89de29b47b71c4b7bd
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>